// client/game/login/renderer/host.ts
import type { BitmapFont } from "../../../rs/font/BitmapFont";
import type { IndexedSprite } from "../../../rs/sprite/IndexedSprite";
import type { LoginScreenAnimation } from "../LoginScreenAnimation";
import type { LoginLayoutConfig, ServerListEntry, World, WorldGridLayout } from "./types";

export type RenderContext = CanvasRenderingContext2D | OffscreenCanvasRenderingContext2D;

export interface LoginRendererHost {
    readonly LOGIN_BOX_X: number;
    readonly LOGIN_BOX_CENTER: number;
    readonly TITLEBOX_Y: number;
    readonly TITLEBOX_FALLBACK_WIDTH: number;
    readonly TITLEBOX_FALLBACK_HEIGHT: number;
    readonly BOTTOM_CONTROLS_RESERVE: number;
    readonly CONTENT_WIDTH: number;
    readonly SCENE_WIDTH: number;
    readonly SCENE_HEIGHT: number;
    titleboxY: number;
    contentOriginY: number;
    contentOriginX: number;
    contentScale: number;
    containerX: number;
    containerWidth: number;
    containerHeight: number;
    xPadding: number;
    canvasWidth: number;
    canvasHeight: number;
    renderScale: number;
    renderOffsetX: number;
    renderOffsetY: number;
    renderSurfaceWidth: number;
    mobileKeyboardFocusActive: boolean;
    mobileKeyboardFocusField: number;
    loginBoxX: number;
    loginBoxCenter: number;
    titleBackgroundImage: ImageBitmap | undefined;
    logoSprite: IndexedSprite | undefined;
    logoImage: HTMLImageElement | undefined;
    logoImageLoaded: boolean;
    titleboxSprite: IndexedSprite | undefined;
    titlebuttonSprite: IndexedSprite | undefined;
    titlebuttonLargeSprite: IndexedSprite | undefined;
    playNowTextSprite: IndexedSprite | undefined;
    runesSprites: IndexedSprite[] | undefined;
    titleMuteSprites: IndexedSprite[] | undefined;
    optionsRadioSprite0: IndexedSprite | undefined;
    optionsRadioSprite2: IndexedSprite | undefined;
    optionsRadioSprite4: IndexedSprite | undefined;
    optionsRadioSprite6: IndexedSprite | undefined;
    worldSelectLeftSprite: IndexedSprite | undefined;
    worldSelectRightSprite: IndexedSprite | undefined;
    worldSelectButtonSprite: IndexedSprite | undefined;
    worldSelectBackSprites: IndexedSprite[] | undefined;
    worldSelectFlagSprites: IndexedSprite[] | undefined;
    worldSelectStarSprites: IndexedSprite[] | undefined;
    worldSelectArrowSprites: IndexedSprite[] | undefined;
    fontBold12: BitmapFont | undefined;
    fontPlain11: BitmapFont | undefined;
    fontPlain12: BitmapFont | undefined;
    cycle: number;
    lastTickTime: number;
    caretBlinkMs: number;
    mouseX: number;
    mouseY: number;
    serverList: ServerListEntry[];
    serverListFetched: boolean;
    probing: boolean;
    probed: boolean;
    worldSortOption: number;
    worldSortDirection: number;
    currentSortedWorlds: World[];
    cachedSortedWorlds: World[] | null;
    cachedSortOption: number;
    cachedSortDirection: number;
    loginScreenRunesAnimation: LoginScreenAnimation | undefined;
    canvas: HTMLCanvasElement | undefined;
    ctx: CanvasRenderingContext2D | undefined;
    spriteCache: WeakMap<IndexedSprite, OffscreenCanvas>;
    textMeasureCache: WeakMap<BitmapFont, Map<string, number>>;
    worldSelectCache: OffscreenCanvas | null;
    worldSelectCacheCtx: OffscreenCanvasRenderingContext2D | null;
    worldSelectCachePage: number;
    worldSelectCacheSortOption: number;
    worldSelectCacheSortDirection: number;
    worldSelectCacheWidth: number;
    worldSelectCacheHeight: number;
    titleCache: OffscreenCanvas | null;
    titleCacheCtx: OffscreenCanvasRenderingContext2D | null;
    titleCacheStateHash: string;
    titleCacheWidth: number;
    titleCacheHeight: number;
    layoutConfig: LoginLayoutConfig;
    scaledLoginBoxX: number;
    scaledLoginBoxCenter: number;
    cachedGridLayout: WorldGridLayout | null;
    cachedGridWorldCount: number;
}